Add files via upload
This file is equivalent to viirs_compact.cdl, only that the interpolation variable has been moved out to the root group.
There is no attribute in the radiance variable referencing the interpolation variable. It is implicitly assumed that all compact variables listed in the interpolation variable are expanded out into the root group. Note that all variables referenced in the interpolation variable now includes the relative path to the tie-point group tp.